Grids: extract AI data tests into a separate file
